+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 
 ### Fixed
 
-*none*
+- properly indent formatted stack trace
 
 ## [0.3.0] - 2022-05-31